Upload OBS docx file step 1
This allows the user to upload a docx file or a zipped docx file and import it into the Door43 namespace. This first step creates a `obs-preview` directory that allows the user to verify the upload was successful. The next step will publish the uploaded stories.

+++ lib/plugins/door43shared/helper.php
@@ -79,6 +79,10 @@ class helper_plugin_door43shared extends DokuWiki_Plugin {
      * @param string $dir
      */
     public function delete_directory_and_files($dir) {
+
+        // don't try to delete it if it doesn't exist
+        if (!file_exists($dir)) return;
+
         foreach(scandir($dir) as $file) {
             if ('.' === $file || '..' === $file) continue;
             if (is_dir("$dir/$file")) $this->delete_directory_and_files("$dir/$file");
